// src/api/generated/models/staffDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * GymFlow API
 * OpenAPI spec version: v1
 */

export interface StaffDto {
  id?: string;
  staffId?: string;
  firstName?: string;
  lastName?: string;
  email?: string;
  phone?: string;
  role?: string;
  salary?: number;
  joinDate?: string;
  status?: string;
  specializations?: string[];
  /** @nullable */
  schedule?: string | null;
  /** @nullable */
  profilePhoto?: string | null;
  createdAt?: string;
}