// src/api/generated/models/createStaffRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * GymFlow API
 * OpenAPI spec version: v1
 */

export interface CreateStaffRequest {
  /**
     * @minLength 1
     * @maxLength 100
     */
  firstName: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  lastName: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  email: string;
  /**
     * @maxLength 20
     * @nullable
     */
  phone?: string | null;
  /** @minLength 1 */
  role: string;
  /** @minimum 0 */
  salary?: number;
  /** @minLength 1 */
  joinDate: string;
  isActive?: boolean;
  /** @nullable */
  specializations?: string[] | null;
  /**
     * @maxLength 200
     * @nullable
     */
  schedule?: string | null;
  /**
     * @maxLength 500
     * @nullable
     */
  profilePhoto?: string | null;
}