// src/api/generated/models/updateStaffRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * GymFlow API
 * OpenAPI spec version: v1
 */

export interface UpdateStaffRequest {
  /**
     * @maxLength 100
     * @nullable
     */
  firstName?: string | null;
  /**
     * @maxLength 100
     * @nullable
     */
  lastName?: string | null;
  /**
     * @maxLength 200
     * @nullable
     */
  email?: string | null;
  /**
     * @maxLength 20
     * @nullable
     */
  phone?: string | null;
  /** @nullable */
  role?: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  salary?: number | null;
  /** @nullable */
  joinDate?: string | null;
  /** @nullable */
  isActive?: boolean | null;
  /** @nullable */
  specializations?: string[] | null;
  /**
     * @maxLength 200
     * @nullable
     */
  schedule?: string | null;
  /**
     * @maxLength 500
     * @nullable
     */
  profilePhoto?: string | null;
}