// src/api/generated/models/problemDetails.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * GymFlow API
 * OpenAPI spec version: v1
 */

export interface ProblemDetails {
  /** @nullable */
  type?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  status?: number | null;
  /** @nullable */
  detail?: string | null;
  /** @nullable */
  instance?: string | null;
  [key: string]: unknown;
}